Mejoras en cierre de órdenes y gastos de importación
- Se reemplaza `ExpenseDate` por `ImportationExpenseDate` en entidad y servicios.
- Se agregan datos descriptivos en DTOs de órdenes (nombres de entidades relacionadas).
- Nueva validación: no se permite cerrar órdenes sin resumen de landed cost.
- Corrección en validación de seguro internacional en landed cost.
- ViewModels de órdenes ahora usan campos opcionales para mejor manejo de nulos.
- Nueva acción y botón para cerrar órdenes desde el listado.
- Vista de landed cost ahora muestra un ViewModel más detallado.
- Ajustes en inyección de dependencias y contexto de base de datos.
- Mejoras en consultas para incluir productos relacionados en detalles de orden.

+++ Application/Services/LandedCostServices/LandedCostService.cs
@@ -48,7 +48,7 @@ public class LandedCostService
         if (order.ImportationExpenses == null || !order.ImportationExpenses.Any(e => e.ExpenseType == ExpenseType.FleteInternacional))
             throw new InvalidOperationException("La orden debe tener un gasto de tipo Flete internacional registrado.");
 
-        if (!order.ImportationExpenses.Any(e => e.ExpenseType == ExpenseType.SeguroInternacional))
+        if (!order.ImportationExpenses.Any(e => (int)e.ExpenseType == (int)ExpenseType.SeguroInternacional))
             throw new InvalidOperationException("La orden debe tener un gasto de tipo Seguro internacional registrado.");
 
         if (order.ImportationExpenses.Any(e => e.DistributionMethod == DistributionMethod.PorVolumen))
